Athena: -KIBL integration with AMQPd and HTTPd daemons. -New Qualifier. -GDL: New UI -GDL: New Javascript engine -GDL: New TicketCart System
Minerve:
in AbstractApp: INTRODUCING generic_post_form_v2: This is an enhanced, varadic version of generic_post_form.
In the long term, v2 will replace the current one.
For now, generic_post_form is a wrapper around v2's varadic calling method.
V2 is dynamic, and can use either an event for a form, or a form itself.
Serialisers now support many-to-many relationships.